component creation
- User Info Box its own component
- Items list its own component

Relevant pages have been updated to just accept the component and pass required props
